+++ apps/extension/src/widgets/layout/ui/index.ts
@@ -1,2 +1,3 @@
 export { default as SidePanelFooter } from "./SidePanelFooter";
+export { default as SidePanelHeader } from "./SidePanelHeader";
 export { default as SidePanelLayout } from "./SidePanelLayout";
